feat: add 멤버 타운 조회

+++ src/main/java/com/themore/moamoatown/member/dto/MemberJobResponseDTO.java
@@ -25,5 +25,5 @@ import lombok.NoArgsConstructor;
 public class MemberJobResponseDTO {
     private String name;
     private String description;
-    private int pay;
+    private Long pay;
 }
